LPS-91859 build rest
