+++ modules/apps/headless/headless-form/headless-form-api/src/main/java/com/liferay/headless/form/resource/v1_0/FormRecordFormResource.java
@@ -35,6 +35,10 @@ public interface FormRecordFormResource {
 			Long formRecordId, FormRecordForm formRecordForm)
 		throws Exception;
 
+	public FormRecord postFormFormRecord(
+			Long formId, FormRecordForm formRecordForm)
+		throws Exception;
+
 	public void setContextCompany(Company contextCompany);
 
 }
