+++ modules/apps/headless/headless-form/headless-form-impl/rest-openapi.yaml
@@ -509,7 +509,7 @@ paths:
                 content:
                     application/json:
                         schema:
-                            $ref: "#/components/schemas/FormRecord"
+                            $ref: "#/components/schemas/FormRecordForm"
             responses:
                 200:
                     content:
@@ -517,7 +517,7 @@ paths:
                             schema:
                                 $ref: "#/components/schemas/FormRecord"
                     description: ""
-            tags: ["FormRecord"]
+            tags: ["FormRecordForm"]
     "/forms/{formId}/form-records/by-latest-draft":
         get:
             operationId: getFormFormRecordByLatestDraft
